Create apiProducts.go to refactor API Product related common methods
