+++ import-export-cli/impl/deleteAPIProduct.go
@@ -19,7 +19,6 @@
 package impl
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
 	"strconv"
@@ -39,7 +38,7 @@ import (
 func DeleteAPIProduct(accessToken, environment, apiProductName, apiProductProvider string) (*resty.Response, error) {
 	deleteEndpoint := utils.GetApiProductListEndpointOfEnv(environment, utils.MainConfigFilePath)
 	deleteEndpoint = utils.AppendSlashToString(deleteEndpoint)
-	apiProductId, err := getAPIProductId(accessToken, environment, apiProductName, apiProductProvider)
+	apiProductId, err := GetAPIProductId(accessToken, environment, apiProductName, apiProductProvider)
 	if err != nil {
 		utils.HandleErrorAndExit("Error while getting API Product Id for deletion ", err)
 	}
@@ -64,55 +63,5 @@ func PrintDeleteAPIProductResponse(resp *resty.Response, err error) {
 		fmt.Println("Error deleting API Product:", err)
 	} else {
 		fmt.Println("API Product deleted successfully!. Status: " + strconv.Itoa(resp.StatusCode()))
-	}
-}
-
-
-// Get the ID of an API Product if available
-// @param accessToken : Access token to call the Publisher Rest API
-// @param environment : Environment where API Product needs to be located
-// @param apiProductName : Name of the API Product
-// @param apiProductProvider : Provider of the API Product
-// @return apiId, error
-func getAPIProductId(accessToken, environment, apiProductName, apiProductProvider string) (string, error) {
-	// Unified Search endpoint from the config file to search API Products
-	unifiedSearchEndpoint := utils.GetUnifiedSearchEndpointOfEnv(environment, utils.MainConfigFilePath)
-
-	// Prepping headers
-	headers := make(map[string]string)
-	headers[utils.HeaderAuthorization] = utils.HeaderValueAuthBearerPrefix + " " + accessToken
-	var queryVal string
-	// TODO Search by version as well when the versioning support has been implemented for API Products
-	queryVal = "type:\"" + utils.DefaultApiProductType + "\" name:\"" + apiProductName + "\""
-	if apiProductProvider != "" {
-		queryVal = queryVal + " provider:\"" + apiProductProvider + "\""
-	}
-	resp, err := utils.InvokeGETRequestWithQueryParam("query", queryVal, unifiedSearchEndpoint, headers)
-	if err != nil {
-		return "", err
-	}
-	if resp.StatusCode() == http.StatusOK || resp.StatusCode() == http.StatusCreated {
-		// 200 OK or 201 Created
-		apiData := &utils.ApiSearch{}
-		data := []byte(resp.Body())
-		err = json.Unmarshal(data, &apiData)
-		if apiData.Count != 0 {
-			apiProductId := apiData.List[0].ID
-			return apiProductId, err
-		}
-		// TODO Print the version as well when the versioning support has been implemented for API Products
-		if apiProductProvider != "" {
-			return "", errors.New("Requested API Product is not available in the Publisher. API Product: " + apiProductName +
-				" Provider: " + apiProductProvider)
-		}
-		return "", errors.New("Requested API Product is not available in the Publisher. API Product: " + apiProductName)
-	} else {
-		utils.Logf("Error: %s\n", resp.Error())
-		utils.Logf("Body: %s\n", resp.Body())
-		if resp.StatusCode() == http.StatusUnauthorized {
-			// 401 Unauthorized
-			return "", fmt.Errorf("Authorization failed while searching API Product: " + apiProductName)
-		}
-		return "", errors.New("Request didn't respond 200 OK for searching API Products. Status: " + resp.Status())
 	}
 }
