+++ import-export-cli/impl/getApiProducts.go
@@ -19,11 +19,8 @@
 package impl
 
 import (
-	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
-	"net/http"
 	"os"
 	"text/template"
 
@@ -91,50 +88,6 @@ func (a *apiProduct) MarshalJSON() ([]byte, error) {
 func GetAPIProductListFromEnv(accessToken, environment, query, limit string) (count int32, apiProducts []utils.APIProduct, err error) {
 	unifiedSearchEndpoint := utils.GetUnifiedSearchEndpointOfEnv(environment, utils.MainConfigFilePath)
 	return GetAPIProductList(accessToken, unifiedSearchEndpoint, query, limit)
-}
-
-// GetAPIProductList
-// @param accessToken : Access Token for the environment
-// @param unifiedSearchEndpoint : Unified Search Endpoint for the environment to retreive API Product list
-// @param query : String to be matched against the API Product names
-// @return count (no. of API Products)
-// @return array of API Product objects
-// @return error
-func GetAPIProductList(accessToken, unifiedSearchEndpoint, query, limit string) (count int32, apiProducts []utils.APIProduct, err error) {
-	// Unified Search endpoint from the config file to search API Products
-	headers := make(map[string]string)
-	headers[utils.HeaderAuthorization] = utils.HeaderValueAuthBearerPrefix + " " + accessToken
-
-	// To filter API Products from unified search
-	unifiedSearchEndpoint += "?query=type:\"" + utils.DefaultApiProductType + "\""
-
-	// Setting up the query parameter and limit parameter
-	if query != "" {
-		unifiedSearchEndpoint += " " + query
-	}
-	if limit != "" {
-		unifiedSearchEndpoint += "&limit=" + limit
-	}
-	utils.Logln(utils.LogPrefixInfo+"URL:", unifiedSearchEndpoint)
-	resp, err := utils.InvokeGETRequest(unifiedSearchEndpoint, headers)
-
-	if err != nil {
-		utils.HandleErrorAndExit("Unable to connect to "+unifiedSearchEndpoint, err)
-	}
-
-	utils.Logln(utils.LogPrefixInfo+"Response:", resp.Status())
-
-	if resp.StatusCode() == http.StatusOK {
-		apiProductListResponse := &utils.APIProductListResponse{}
-		unmarshalError := json.Unmarshal([]byte(resp.Body()), &apiProductListResponse)
-
-		if unmarshalError != nil {
-			utils.HandleErrorAndExit(utils.LogPrefixError+"invalid JSON response", unmarshalError)
-		}
-		return apiProductListResponse.Count, apiProductListResponse.List, nil
-	} else {
-		return 0, nil, errors.New(string(resp.Body()))
-	}
 }
 
 // PrintAPIProducts
